Rename `convertScoreToColorAndEmoji.js` to `index.js` and edit references to it
